[FIX] Fixes crash "dir doesn't exist' on fs.rmdirSync

+++ npm_module/scripts/build.js
@@ -33,7 +33,7 @@ function update_entries() {
 }
 
 function clean() {
-	fs.rmdirSync(options.outRoot, {recursive: true});
+	if (fs.existsSync(options.outRoot)) fs.rmdirSync(options.outRoot, {recursive: true});
 }
 
 function watch() {
